Update 10/18/25 3:47AM
I added the target points necessary for the enemies to patrol the area around each room.

I added the end door, where it announces to the player that they've won.

I added skins to both my players character model, and the enemy models to make them look the part.

I also added a crouch and dash that the players will be able to use to navigate around the rooms.